python/mcp/pyproject.toml
[project]
name = "mcp-pro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.49.0",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.5.0",
    "openai>=1.68.2",
    "python-dotenv>=1.1.0",
]
